{kirkstone} added pkgconfig class to ros_distro_humble
This is a rather awkward hack. libyaml-vendor deploys a Findyaml.cmake
that uses pkgconfig to find libyaml and adds it to the ${TARGET}. Seems
like pkgconfig was implicitly pulled in with honister, this is no longer
the case. The Findyaml.cmake silently fails and we get linking errors
because of missing libyaml symbols. This is in rcl-yaml-param-parser,
which is included in rclcpp, and that gets included a lot of course.

Instead of adding tons of bbappends with "inherit pkgconfig" in it, we
include it into the bbclass.

Signed-off-by: Matthias Schoepfer <m.schoepfer@rethinkrobotics.com>